// src/core/api/model/patientsControllerFindAllSortBy.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Dentizy API
 * API Documentation untuk Sistem Manajemen Klinik Gigi
 * OpenAPI spec version: 1.0
 */

export type PatientsControllerFindAllSortBy = typeof PatientsControllerFindAllSortBy[keyof typeof PatientsControllerFindAllSortBy];


export const PatientsControllerFindAllSortBy = {
  nama_lengkap: 'nama_lengkap',
  nomor_rekam_medis: 'nomor_rekam_medis',
  tanggal_lahir: 'tanggal_lahir',
  created_at: 'created_at',
  umur: 'umur',
} as const;