// src/core/api/model/usersControllerFindAllRole.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Dentizy API
 * API Documentation untuk Sistem Manajemen Klinik Gigi
 * OpenAPI spec version: 1.0
 */

export type UsersControllerFindAllRole = typeof UsersControllerFindAllRole[keyof typeof UsersControllerFindAllRole];


export const UsersControllerFindAllRole = {
  dokter: 'dokter',
  staf: 'staf',
  kepala_klinik: 'kepala_klinik',
} as const;